Deploy examples

+++ lib/onnx/nodes/batchNormalization.ts
@@ -11,6 +11,9 @@ export class BatchNormalizationNode extends OnnxNode {
 
   private epsTensor: Tensor;
 
+  private scale?: Tensor;
+  private bias?: Tensor;
+
   constructor(attributes: Attributes, inputs: string[], outputs: string[], constants: Constants, onnxVersion: number) {
     super(attributes, inputs, outputs, constants, onnxVersion);
 
@@ -22,55 +25,72 @@ export class BatchNormalizationNode extends OnnxNode {
     //TODO: Handle lower onnxversions here
   }
 
+  initialize(resolveConstant: (name: string) => Tensor) {
+    const scale = resolveConstant(this.inputs[1]);
+    const B = resolveConstant(this.inputs[2]);
+    const mean = resolveConstant(this.inputs[3]);
+    const variance = resolveConstant(this.inputs[4]);
+
+    if (scale !== undefined && B !== undefined && mean !== undefined && variance !== undefined) {
+      const varSqrt = variance.add(this.epsTensor).sqrt();
+
+      this.scale = scale.divide(varSqrt);
+      this.bias = B.subtract(mean.multiply(this.scale));
+
+      varSqrt.delete();
+    }
+  }
+
   async forward(inputs: Tensor[]): Promise<Tensor[]> {
     const x = inputs[0];
-    let scale = inputs[1];
-    let B = inputs[2];
-    let mean = inputs[3];
-    let variance = inputs[4];
 
-    //TODO: Handle lower onnx versions here
+    if (this.scale !== undefined) {
+      const scaled = x.multiply(this.scale);
+      const result = scaled.add(this.bias);
+      scaled.delete();
 
-    const C = scale.getShape()[0];
+      return [result];
+    } else {
+      let scale = inputs[1];
+      let B = inputs[2];
+      let mean = inputs[3];
+      let variance = inputs[4];
 
-    const newShape = [1,C,...new Array(x.getShape().length - 2).fill(1)];
+      //TODO: Handle lower onnx versions here
 
-    scale = scale.reshape(newShape);
-    B = B.reshape(newShape);
-    mean = mean.reshape(newShape);
-    variance = variance.reshape(newShape);
+      const C = scale.getShape()[0];
 
-    glContext.flush();
-    const varEps = variance.add(this.epsTensor);
-    glContext.flush();
-    const varEpsSqrt = varEps.sqrt();
-    const xmean = x.subtract(mean);
-    glContext.flush();
-    const normalized = xmean.divide(varEpsSqrt);
+      const newShape = [1,C,...new Array(x.getShape().length - 2).fill(1)];
 
-    glContext.flush();
-    const scaled = normalized.multiply(scale);
-    glContext.flush();
-    const result = scaled.add(B);
+      scale = scale.reshape(newShape, false);
+      B = B.reshape(newShape, false);
+      mean = mean.reshape(newShape, false);
+      variance = variance.reshape(newShape, false);
 
-    varEps.delete();
-    varEpsSqrt.delete();
-    xmean.delete();
-    normalized.delete();
-    scaled.delete();
+      const result = x.normalize(mean, variance, this.epsilon, scale, B);
 
-    // TODO: Replace this with native batchnorm implementation
-
-    return [result];
+      return [result];
+    }
   }
 
   async toCPU() {
-    this.epsTensor = await toCPU(this.epsTensor);
+    if (this.scale !== undefined) {
+      this.scale = await toCPU(this.scale);
+      this.bias = await toCPU(this.bias);
+    }
   }
+
   async toWASM() {
-    this.epsTensor = await toWASM(this.epsTensor);
+    if (this.scale !== undefined) {
+      this.scale = await toWASM(this.scale);
+      this.bias = await toWASM(this.bias);
+    }
   }
+
   async toGPU() {
-    this.epsTensor = await toGPU(this.epsTensor);
+    if (this.scale !== undefined) {
+      this.scale = await toGPU(this.scale);
+      this.bias = await toGPU(this.bias);
+    }
   }
 }
